+++ index.js
@@ -5,6 +5,8 @@ const loginCommon = require('./lib/common/login.js');
 const logoutCommon = require('./lib/common/logout.js');
 const tenant = require('./lib/common/tenant.js');
 const profile = require('./lib/auth0/profile.js');
+const getJWT = require('./lib/utils/get-jwt.js');
+const settings = require('./lib/common/settings.js');
 
 const privateVars = new WeakMap();
 
@@ -49,10 +51,33 @@ class BlinkMobileIdentity {
 
   /**
    * Get the Auth0 profile for the current user.
+   * @param {String} [accessToken] - The access token generated after a successful login. If not passed, will attempt to get the access token from the file system.
    * @returns {Object} The Auth0 profile.
    */
-  getProfile () {
-    return profile.getByClient(privateVars.get(this).clientName);
+  getProfile (accessToken) {
+    if (accessToken) {
+      return profile.getByJWT(accessToken);
+    } else {
+      return profile.getByClient(privateVars.get(this).clientName);
+    }
+  }
+
+  /**
+   * Get access token generated after a successful login
+   * @function getAccessToken
+   * @returns {String} The access token generated after a successful login.
+   */
+  getAccessToken () {
+    return getJWT();
+  }
+
+  /**
+   * Get settings scoped to a BlinkMobile service.
+   * @function getServiceSettings
+   * @returns {Object} The settings.
+   */
+  getServiceSettings () {
+    return settings(privateVars.get(this).clientName);
   }
 
   /**